Move all of the options to the config.go file.
It feels cleaner to have all of these in one file.
