+++ config.go
@@ -1,8 +1,12 @@
 package unleash
 
 import (
-	"github.com/unleash/unleash-client-go/strategy"
+	"net/http"
+	"net/url"
 	"time"
+
+	"github.com/unleash/unleash-client-go/context"
+	"github.com/unleash/unleash-client-go/strategy"
 )
 
 type configOption struct {
@@ -64,4 +68,44 @@ func WithStrategies(strategies ...strategy.Strategy) ConfigOption {
 	return func(o *configOption) {
 		o.strategies = strategies
 	}
+}
+
+type featureOption struct {
+	fallback *bool
+	ctx      *context.Context
+}
+
+type FeatureOption func(*featureOption)
+
+func WithFallback(fallback bool) FeatureOption {
+	return func(opts *featureOption) {
+		opts.fallback = &fallback
+	}
+}
+
+func WithContext(ctx context.Context) FeatureOption {
+	return func(opts *featureOption) {
+		opts.ctx = &ctx
+	}
+}
+
+type RepositoryOptions struct {
+	AppName         string
+	InstanceId      string
+	Url             url.URL
+	BackupPath      string
+	RefreshInterval time.Duration
+	Storage         Storage
+	HttpClient      *http.Client
+}
+
+type MetricsOptions struct {
+	AppName         string
+	InstanceID      string
+	Url             url.URL
+	Strategies      []string
+	MetricsInterval time.Duration
+	DisableMetrics  bool
+	BucketInterval  time.Duration
+	HttpClient      *http.Client
 }
